+++ src/main/java/org/ton/exceptions/WrongSeqnoException.java
@@ -1,0 +1,7 @@
+package org.ton.exceptions;
+
+public class WrongSeqnoException extends Exception {
+    public WrongSeqnoException(String errorMessage) {
+        super(errorMessage);
+    }
+}